unit-tests-service/requirements.txt
Flask==2.3.3
python-dotenv
together
requests 
fastapi
uvicorn[standard]

